/login: Delete old unused <CreateAccount/> component
